TD_cycle_type - now with classical TD cycles - half completed (qs style TD cycles work)

+++ hqf/hqf_fec_prepare_one_msp.sh
@@ -1,9 +1,7 @@
 #!/usr/bin/env bash 
 
 # Usage infomation
-usage="Usage: hqf_ce_prepare_one_fes.py
-
-<msp_name> <subsystem> <pipeline_type>
+usage="Usage: hqf_fec_prepare_one_msp.sh <system 1 basename> <system 2 basename> <subsystem>
 
 The script has to be run in the root folder."
 
@@ -64,6 +62,21 @@ stride_fec=$(grep -m 1 "^stride_fec=" input-files/config.txt | awk -F '=' '{prin
 # Printing some information
 echo -e "\n *** Preparing the FEC between the systems ${system_1_basename} and ${system_2_basename}"
 
+# Checking if the md_folder exists
+if [ ! -d "${md_folder}" ]; then
+    echo -e "\nError: The folder ${md_folder} does not exist. Exiting\n\n" 1>&2
+    exit 1
+fi
+
+# Checking if the ce_folder exists
+if [ ! -d "${ce_folder}" ]; then
+    echo "\nError: The folder ${ce_folder} does not exist. Exiting\n\n" 1>&2
+    exit 1
+fi
+
+# Checking if the ce_folder exists
+md_folder="md/${msp_name}/${subsystem}"
+
 # Creating required folders
 if [ -d ${fec_folder} ]; then         
     rm -r ${fec_folder}
@@ -74,14 +87,15 @@ mkdir -p ${fec_folder}
 while read line; do 
     md_folder_1="$(echo -n ${line} | awk '{print $1}')"
     md_folder_2="$(echo -n ${line} | awk '{print $2}')"    
-    crosseval_folder_fw="${md_folder_2}-${md_folder_1}"     # md folder1 (positions) is evaluated at folder two's potential: potentialfolder-positionfolder
-    crosseval_folder_bw="${md_folder_1}-${md_folder_2}"     # Opposite of fw
+    crosseval_folder_fw="${md_folder_1}-${md_folder_2}"     # md folder1 (positions) is evaluated at folder two's potential: samplingfolder-potentialfolder
+    crosseval_folder_bw="${md_folder_2}-${md_folder_1}"     # Opposite of fw
     
     # Variables 
     stride_ipi_properties=$(grep "potential" ${md_folder}/${md_folder_1}/ipi/ipi.in.md.xml | tr -s " " "\n" | grep "stride" | awk -F '=' '{print $2}' | tr -d '"')
     stride_ipi_trajectory=$(grep "<checkpoint" ${md_folder}/${md_folder_1}/ipi/ipi.in.md.xml | tr -s " " "\n" | grep "stride" | awk -F '=' '{print $2}' | tr -d '"')
     stride_ipi=$((stride_ipi_trajectory  / stride_ipi_properties))
     stride_fec=$(grep -m 1 "^stride_fec=" input-files/config.txt | awk -F '=' '{print $2}')
+    stride_ce=$(grep -m 1 "^stride_ce=" input-files/config.txt | awk -F '=' '{print $2}')
     TD_window=${md_folder_1}-${md_folder_2}
 
     # Checking if the ipi-input-file <variables stride_ipi_properties> and <stride_ipi_trajectory> are compatible
@@ -100,40 +114,36 @@ while read line; do
 
     # Loop for the each snapshot pair
     snapshotCount=$(ls ${ce_folder}/${crosseval_folder_fw} | wc -l)
-    for snapshot_ID in $(seq 0 $((snapshotCount -1))); do
+    for i in $(seq 0 $((snapshotCount -1))); do
+        snapshot_ID=$((i*stride_ce))
         snapshot_folder=snapshot-${snapshot_ID}
 
         # Foward direction
-        # Extracting the free energies of system 1 evaluated at system 2
-        energy_fw_U2_U1="$(grep -v "^#" ${ce_folder}/${crosseval_folder_fw}/${snapshot_folder}/ipi/ipi.out.properties | awk '{print $4}')"
         # Extracting the free energies of system 1 evaluated at system 1
         energy_fw_U1_U1="$(grep -v "^#" ${md_folder}/${md_folder_1}/ipi/ipi.out.properties | awk '{print $4}' | tail -n+$((snapshot_ID +1)) | head -n 1)"
+        # Extracting the free energies of system 1 evaluated at system 2
+        energy_fw_U1_U2="$(grep -v "^#" ${ce_folder}/${crosseval_folder_fw}/${snapshot_folder}/ipi/ipi.out.properties | awk '{print $4}')"
 
-        if [[ "${energy_fw_U2_U1}" == *[0-9]* ]] && [[ "${energy_fw_U1_U1}" == *[0-9]* ]]; then
-            echo "${energy_fw_U2_U1}" >>  ${fec_folder}/${TD_window}/U2_U1
+        if [[ "${energy_fw_U1_U1}" == *[0-9]* ]] && [[ "${energy_fw_U1_U2}" == *[0-9]* ]]; then
             echo "${energy_fw_U1_U1}" >>  ${fec_folder}/${TD_window}/U1_U1
+            echo "${energy_fw_U1_U2}" >>  ${fec_folder}/${TD_window}/U1_U2
         fi
     done
     snapshotCount=$(ls ${ce_folder}/${crosseval_folder_bw} | wc -l)
-    for snapshot_ID in $(seq 0 $((snapshotCount -1))); do
+    for i in $(seq 0 $((snapshotCount -1))); do
+        snapshot_ID=$((i*stride_ce))
         snapshot_folder=snapshot-${snapshot_ID}
 
         # Backward direction
         # Extracting the free energies of system 2 evaluated at system 1
-        energy_bw_U1_U2="$(grep -v "^#" ${ce_folder}/${crosseval_folder_bw}/${snapshot_folder}/ipi/ipi.out.properties | awk '{print $4}')"
+        energy_bw_U2_U1="$(grep -v "^#" ${ce_folder}/${crosseval_folder_bw}/${snapshot_folder}/ipi/ipi.out.properties | awk '{print $4}')"
         # Extracting the free energies of system 2 evaluated at system 2
         energy_bw_U2_U2="$(grep -v "^#" ${md_folder}/${md_folder_2}/ipi/ipi.out.properties | awk '{print $4}' | tail -n+$((snapshot_ID +1)) | head -n 1)"
 
         # Checking if the energies were computed and extracted successfully, which is usually the case if there is a number in the value
-        #if [[ "${energy_bw_U1_U2}" == *[0-9]* ]] && [[ "${energy_bw_U2_U2}" == *[0-9]* ]] && [[ "${energy_fw_U2_U1}" == *[0-9]* ]] && [[ "${energy_fw_U1_U1}" == *[0-9]* ]]; then
-        #    echo "${energy_fw_U2_U1}" >>  ${fec_folder}/${TD_window}/U2_U1
-        #    echo "${energy_fw_U1_U1}" >>  ${fec_folder}/${TD_window}/U1_U1
-        #    echo "${energy_bw_U1_U2}" >>  ${fec_folder}/${TD_window}/U1_U2
-        #    echo "${energy_bw_U2_U2}" >>  ${fec_folder}/${TD_window}/U2_U2
-        #fi
-        if [[ "${energy_bw_U1_U2}" == *[0-9]* ]] && [[ "${energy_bw_U2_U2}" == *[0-9]* ]]; then
-            echo "${energy_bw_U1_U2}" >>  ${fec_folder}/${TD_window}/U1_U2
-            echo "${energy_bw_U2_U2}" >>  ${fec_folder}/${TD_window}/U2_U2
+        if [[ "${energy_bw_U2_U1}" == *[0-9]* ]] && [[ "${energy_bw_U2_U2}" == *[0-9]* ]]; then
+            echo "${energy_bw_U2_U1}" >> ${fec_folder}/${TD_window}/U2_U1
+            echo "${energy_bw_U2_U2}" >> ${fec_folder}/${TD_window}/U2_U2
         fi
     done
     
@@ -144,10 +154,16 @@ while read line; do
     sed -n "0~${stride_fec}p" ${fec_folder}/${TD_window}/U2_U2 > ${fec_folder}/${TD_window}/U2_U2_stride${stride_fec}
     
     # Computing the free energy differences delta U (mainly for Mobley pymbar code)
-    paste ${fec_folder}/${TD_window}/U1_U2 ${fec_folder}/${TD_window}/U2_U2 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U2-U2_U2
-    paste ${fec_folder}/${TD_window}/U2_U1 ${fec_folder}/${TD_window}/U1_U1 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U1-U1_U1
-    paste ${fec_folder}/${TD_window}/U1_U2_stride${stride_fec} ${fec_folder}/${TD_window}/U2_U2_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U2-U2_U2_stride${stride_fec}
-    paste ${fec_folder}/${TD_window}/U2_U1_stride${stride_fec} ${fec_folder}/${TD_window}/U1_U1_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U1-U1_U1_stride${stride_fec}
+    paste ${fec_folder}/${TD_window}/U1_U1 ${fec_folder}/${TD_window}/U1_U2 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U2-U1_U1 # (U2-U1)_1 -> for out IP method implementation, and for original BAR equation of Bennett (denominator) and our implementation
+    paste ${fec_folder}/${TD_window}/U2_U1 ${fec_folder}/${TD_window}/U2_U2 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U2-U2_U1 # (U2-U1)_2 -> for our IP method implemenation
+    paste ${fec_folder}/${TD_window}/U2_U2 ${fec_folder}/${TD_window}/U2_U1 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U1-U2_U2 # for our original BAR equation of Bennett (nominator) implemenation
+    paste ${fec_folder}/${TD_window}/U1_U2 ${fec_folder}/${TD_window}/U1_U1 | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U1-U1_U2 # for our original BAR equation of Bennett (nominator) implemenation
+    
+    
+    paste ${fec_folder}/${TD_window}/U2_U1_stride${stride_fec} ${fec_folder}/${TD_window}/U2_U2_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U1-U2_U2_stride${stride_fec}
+    paste ${fec_folder}/${TD_window}/U1_U2_stride${stride_fec} ${fec_folder}/${TD_window}/U1_U1_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U2-U1_U1_stride${stride_fec}
+    paste ${fec_folder}/${TD_window}/U2_U2_stride${stride_fec} ${fec_folder}/${TD_window}/U2_U1_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U2_U2-U2_U1_stride${stride_fec}
+    paste ${fec_folder}/${TD_window}/U1_U1_stride${stride_fec} ${fec_folder}/${TD_window}/U1_U1_stride${stride_fec} | awk '{print $2 - $1}' > ${fec_folder}/${TD_window}/U1_U1-U1_U2_stride${stride_fec}
     
     # Preparing the C-values
     hqh_fec_prepare_cvalues.py ${c_values_min} ${c_values_max} ${c_values_count} > ${fec_folder}/${TD_window}/C-values
